Fix handling of missing metadata file
(cherry picked from commit 2865959ddf1e92a5c2d89e91bc7069a7eafd29e2)

+++ urbackupserver/IncrFileBackup.cpp
@@ -736,45 +736,48 @@ bool IncrFileBackup::doFileBackup()
 							{
 								if(!Server->deleteFile(os_file_prefix(metadata_fn)))
 								{
-									if (sym_target == extra_params.end())
+									if (sym_target == extra_params.end() )
 									{
-										//Directory symlink to directory
-										if (!os_remove_symlink_dir(os_file_prefix(backuppath + local_curr_os_path)) )
+										if(os_get_file_type(os_file_prefix(backuppath + local_curr_os_path)) & EFileType_Symlink )
 										{
-											ServerLogger::Log(logid, "Could not remove symbolic link at \"" + backuppath + local_curr_os_path + "\" (2). " + systemErrorInfo(), LL_ERROR);
-											c_has_error = true;
-											break;
-										}
-
-										if (!os_create_dir(os_file_prefix(backuppath + local_curr_os_path)))
-										{
-											if (!os_directory_exists(os_file_prefix(backuppath + local_curr_os_path)))
+											//Directory symlink to directory
+											if (!os_remove_symlink_dir(os_file_prefix(backuppath + local_curr_os_path)) )
 											{
-												ServerLogger::Log(logid, "Creating directory  \"" + backuppath + local_curr_os_path + "\" failed. - " + systemErrorInfo(), LL_ERROR);
+												ServerLogger::Log(logid, "Could not remove symbolic link at \"" + backuppath + local_curr_os_path + "\" (2). " + systemErrorInfo(), LL_ERROR);
 												c_has_error = true;
 												break;
 											}
-											else
+
+											if (!os_create_dir(os_file_prefix(backuppath + local_curr_os_path)))
 											{
-												ServerLogger::Log(logid, "Directory \"" + backuppath + local_curr_os_path + "\" does already exist.", LL_WARNING);
+												if (!os_directory_exists(os_file_prefix(backuppath + local_curr_os_path)))
+												{
+													ServerLogger::Log(logid, "Creating directory  \"" + backuppath + local_curr_os_path + "\" failed. - " + systemErrorInfo(), LL_ERROR);
+													c_has_error = true;
+													break;
+												}
+												else
+												{
+													ServerLogger::Log(logid, "Directory \"" + backuppath + local_curr_os_path + "\" does already exist.", LL_WARNING);
+												}
 											}
-										}
 
-										std::string metadata_fn_curr = backuppath_hashes + convertToOSPathFromFileClient(orig_curr_os_path + "/" + escape_metadata_fn(cf.name));
-										if(!Server->deleteFile(os_file_prefix(metadata_fn_curr)))
-										{
-											ServerLogger::Log(logid, "Error deleting metadata file \"" + metadata_fn_curr + "\". " + os_last_error_str(), LL_ERROR);
-											c_has_error = true;
-										}
+											std::string metadata_fn_curr = backuppath_hashes + convertToOSPathFromFileClient(orig_curr_os_path + "/" + escape_metadata_fn(cf.name));
+											if(!Server->deleteFile(os_file_prefix(metadata_fn_curr)))
+											{
+												ServerLogger::Log(logid, "Error deleting metadata file \"" + metadata_fn_curr + "\". " + os_last_error_str(), LL_WARNING);
+											}
 
-										metadata_srcpath = last_backuppath_hashes + convertToOSPathFromFileClient(orig_curr_os_path + "/" + escape_metadata_fn(cf.name));
+											metadata_srcpath = last_backuppath_hashes + convertToOSPathFromFileClient(orig_curr_os_path + "/" + escape_metadata_fn(cf.name));
+										}
 									}
-									else if( !Server->deleteFile(os_file_prefix(metadata_fn + os_file_sep() + metadata_dir_fn))
-											|| !os_remove_dir(os_file_prefix(metadata_fn)) )
+									else if(os_get_file_type(os_file_prefix(backuppath + local_curr_os_path)) & EFileType_Symlink == 0 )
 									{
-										ServerLogger::Log(logid, "Error deleting metadata file \"" + metadata_fn + "\". " + os_last_error_str(), LL_ERROR);
-										c_has_error = true;
-										break;
+										if( !Server->deleteFile(os_file_prefix(metadata_fn + os_file_sep() + metadata_dir_fn))
+											|| !os_remove_dir(os_file_prefix(metadata_fn)) )
+										{
+											ServerLogger::Log(logid, "Error deleting metadata file \"" + metadata_fn + os_file_sep() + metadata_dir_fn + "\". " + os_last_error_str(), LL_WARNING);
+										}
 									}
 								}
 							}
